requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlmodel
psycopg[binary]
python-dotenv
